[UIKit] Implemented shared rendering implementation for UIKit based platform
