+++ platform/iphone/godot_view.mm
@@ -37,9 +37,7 @@
 #import <OpenGLES/EAGLDrawable.h>
 #import <QuartzCore/QuartzCore.h>
 
-#import "display_layer.h"
 #import "godot_view_gesture_recognizer.h"
-#import "godot_view_renderer.h"
 
 #import <CoreMotion/CoreMotion.h>
 
@@ -49,17 +47,6 @@ static const int max_touches = 8;
 	UITouch *godot_touches[max_touches];
 }
 
-@property(assign, nonatomic) BOOL isActive;
-
-// CADisplayLink available on 3.1+ synchronizes the animation timer & drawing with the refresh rate of the display, only supports animation intervals of 1/60 1/30 & 1/15
-@property(strong, nonatomic) CADisplayLink *displayLink;
-
-// An animation timer that, when animation is started, will periodically call -drawView at the given rate.
-// Only used if CADisplayLink is not
-@property(strong, nonatomic) NSTimer *animationTimer;
-
-@property(strong, nonatomic) CALayer<DisplayLayer> *renderingLayer;
-
 @property(strong, nonatomic) CMMotionManager *motionManager;
 
 @property(strong, nonatomic) GodotViewGestureRecognizer *delayGestureRecognizer;
@@ -67,26 +54,6 @@ static const int max_touches = 8;
 @end
 
 @implementation GodotView
-
-// Implement this to override the default layer class (which is [CALayer class]).
-// We do this so that our view will be backed by a layer that is capable of OpenGL ES rendering.
-- (CALayer<DisplayLayer> *)initializeRendering {
-	if (self.renderingLayer) {
-		return self.renderingLayer;
-	}
-
-	CALayer<DisplayLayer> *layer = [GodotOpenGLLayer layer];
-
-	layer.frame = self.bounds;
-	layer.contentsScale = self.contentScaleFactor;
-
-	[self.layer addSublayer:layer];
-	self.renderingLayer = layer;
-
-	[layer initializeDisplayLayer];
-
-	return self.renderingLayer;
-}
 
 - (instancetype)initWithCoder:(NSCoder *)coder {
 	self = [super initWithCoder:coder];
@@ -110,29 +77,9 @@ static const int max_touches = 8;
 
 // Stop animating and release resources when they are no longer needed.
 - (void)dealloc {
-	[self stopRendering];
-
-	self.renderer = nil;
-	self.delegate = nil;
-
-	if (self.renderingLayer) {
-		[self.renderingLayer removeFromSuperlayer];
-		self.renderingLayer = nil;
-	}
-
 	if (self.motionManager) {
 		[self.motionManager stopDeviceMotionUpdates];
 		self.motionManager = nil;
-	}
-
-	if (self.displayLink) {
-		[self.displayLink invalidate];
-		self.displayLink = nil;
-	}
-
-	if (self.animationTimer) {
-		[self.animationTimer invalidate];
-		self.animationTimer = nil;
 	}
 
 	if (self.delayGestureRecognizer) {
@@ -141,8 +88,6 @@ static const int max_touches = 8;
 }
 
 - (void)godot_commonInit {
-	self.contentScaleFactor = [UIScreen mainScreen].nativeScale;
-
 	[self initTouches];
 
 	self.multipleTouchEnabled = YES;
@@ -164,122 +109,16 @@ static const int max_touches = 8;
 	[self addGestureRecognizer:self.delayGestureRecognizer];
 }
 
-- (void)startRendering {
-	if (self.isActive) {
-		return;
-	}
+- (void)drawView {
+	[super drawView];
 
-	self.isActive = YES;
-
-	printf("start animation!\n");
-
-	if (self.useCADisplayLink) {
-		self.displayLink = [CADisplayLink displayLinkWithTarget:self selector:@selector(drawView)];
-
-		// Approximate frame rate
-		// assumes device refreshes at 60 fps
-		int displayFPS = (NSInteger)(1.0 / self.renderingInterval);
-
-		self.displayLink.preferredFramesPerSecond = displayFPS;
-
-		// Setup DisplayLink in main thread
-		[self.displayLink addToRunLoop:[NSRunLoop currentRunLoop] forMode:NSRunLoopCommonModes];
-	} else {
-		self.animationTimer = [NSTimer scheduledTimerWithTimeInterval:self.renderingInterval target:self selector:@selector(drawView) userInfo:nil repeats:YES];
-	}
+	[self handleMotion];
 }
 
 - (void)stopRendering {
-	if (!self.isActive) {
-		return;
-	}
-
-	self.isActive = NO;
-
-	printf("******** stop animation!\n");
-
-	if (self.useCADisplayLink) {
-		[self.displayLink invalidate];
-		self.displayLink = nil;
-	} else {
-		[self.animationTimer invalidate];
-		self.animationTimer = nil;
-	}
+	[super stopRendering];
 
 	[self clearTouches];
-}
-
-// Updates the OpenGL view when the timer fires
-- (void)drawView {
-	if (!self.isActive) {
-		printf("draw view not active!\n");
-		return;
-	}
-
-	if (self.useCADisplayLink) {
-		// Pause the CADisplayLink to avoid recursion
-		[self.displayLink setPaused:YES];
-
-		// Process all input events
-		while (CFRunLoopRunInMode(kCFRunLoopDefaultMode, 0.0, TRUE) == kCFRunLoopRunHandledSource)
-			;
-
-		// We are good to go, resume the CADisplayLink
-		[self.displayLink setPaused:NO];
-	}
-
-	[self.renderingLayer startRenderDisplayLayer];
-
-	if (!self.renderer) {
-		return;
-	}
-
-	if ([self.renderer setupView:self]) {
-		return;
-	}
-
-	if (self.delegate) {
-		BOOL delegateFinishedSetup = [self.delegate godotViewFinishedSetup:self];
-
-		if (!delegateFinishedSetup) {
-			return;
-		}
-	}
-
-	[self handleMotion];
-	[self.renderer renderOnView:self];
-
-	[self.renderingLayer stopRenderDisplayLayer];
-}
-
-- (BOOL)canRender {
-	if (self.useCADisplayLink) {
-		return self.displayLink != nil;
-	} else {
-		return self.animationTimer != nil;
-	}
-}
-
-- (void)setRenderingInterval:(NSTimeInterval)renderingInterval {
-	_renderingInterval = renderingInterval;
-
-	if (self.canRender) {
-		[self stopRendering];
-		[self startRendering];
-	}
-}
-
-// If our view is resized, we'll be asked to layout subviews.
-// This is the perfect opportunity to also update the framebuffer so that it is
-// the same size as our display area.
-
-- (void)layoutSubviews {
-	if (self.renderingLayer) {
-		self.renderingLayer.frame = self.bounds;
-		[self.renderingLayer layoutDisplayLayer];
-	}
-
-	[super layoutSubviews];
 }
 
 // MARK: - Input
