+++ platform/iphone/godot_view_controller.mm
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  view_controller.mm                                                   */
+/*  godot_view_controller.mm                                             */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,7 +28,7 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#import "view_controller.h"
+#import "godot_view_controller.h"
 
 #include "core/project_settings.h"
 #import "godot_view.h"
@@ -37,7 +37,7 @@
 #import "native_video_view.h"
 #include "os_iphone.h"
 
-@interface ViewController () <GodotViewDelegate>
+@interface GodotViewController ()
 
 @property(strong, nonatomic) GodotViewRenderer *renderer;
 @property(strong, nonatomic) GodotNativeVideoView *videoView;
@@ -47,7 +47,7 @@
 
 @end
 
-@implementation ViewController
+@implementation GodotViewController
 
 - (GodotView *)godotView {
 	return (GodotView *)self.view;
@@ -90,16 +90,10 @@
 	// Initialize view controller values.
 }
 
-- (void)didReceiveMemoryWarning {
-	[super didReceiveMemoryWarning];
-	printf("*********** did receive memory warning!\n");
-};
-
 - (void)viewDidLoad {
 	[super viewDidLoad];
 
 	[self observeKeyboard];
-	[self displayLoadingOverlay];
 
 	if (@available(iOS 11.0, *)) {
 		[self setNeedsUpdateOfScreenEdgesDeferringSystemGestures];
@@ -124,31 +118,6 @@
 				 object:nil];
 }
 
-- (void)displayLoadingOverlay {
-	NSBundle *bundle = [NSBundle mainBundle];
-	NSString *storyboardName = @"Launch Screen";
-
-	if ([bundle pathForResource:storyboardName ofType:@"storyboardc"] == nil) {
-		return;
-	}
-
-	UIStoryboard *launchStoryboard = [UIStoryboard storyboardWithName:storyboardName bundle:bundle];
-
-	UIViewController *controller = [launchStoryboard instantiateInitialViewController];
-	self.godotLoadingOverlay = controller.view;
-	self.godotLoadingOverlay.frame = self.view.bounds;
-	self.godotLoadingOverlay.autoresizingMask = UIViewAutoresizingFlexibleHeight | UIViewAutoresizingFlexibleWidth;
-
-	[self.view addSubview:self.godotLoadingOverlay];
-}
-
-- (BOOL)godotViewFinishedSetup:(GodotView *)view {
-	[self.godotLoadingOverlay removeFromSuperview];
-	self.godotLoadingOverlay = nil;
-
-	return YES;
-}
-
 - (void)dealloc {
 	[self.videoView stopVideo];
 	self.videoView = nil;
@@ -156,11 +125,6 @@
 	self.keyboardView = nil;
 
 	self.renderer = nil;
-
-	if (self.godotLoadingOverlay) {
-		[self.godotLoadingOverlay removeFromSuperview];
-		self.godotLoadingOverlay = nil;
-	}
 
 	[[NSNotificationCenter defaultCenter] removeObserver:self];
 }
