+++ platform/iphone/platform_config.h
@@ -28,18 +28,4 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include <alloca.h>
-
-#define GLES2_INCLUDE_H <ES2/gl.h>
-#define GLES3_INCLUDE_H <ES3/gl.h>
-
-#define PLATFORM_REFCOUNT
-
-#define PTHREAD_RENAME_SELF
-
-#define _weakify(var) __weak typeof(var) GDWeak_##var = var;
-#define _strongify(var)                                      \
-	_Pragma("clang diagnostic push")                         \
-			_Pragma("clang diagnostic ignored \"-Wshadow\"") \
-					__strong typeof(var) var = GDWeak_##var; \
-	_Pragma("clang diagnostic pop")
+#import "platform/uikit/uikit_platform_config.h"
