+++ platform/uikit/uikit_view_controller.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  godot_view_renderer.mm                                               */
+/*  uikit_view_controller.h                                              */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,29 +28,14 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#import "godot_view_renderer.h"
-
-#include "os_iphone.h"
-
 #import <UIKit/UIKit.h>
 
-@interface GodotViewRenderer ()
+#import "uikit_view.h"
 
-@end
+@class UIKitView;
 
-@implementation GodotViewRenderer
+@interface UIKitViewController : UIViewController <UIKitViewDelegate>
 
-- (BOOL)startUIKitPlatform {
-	OSIPhone::get_singleton()->start();
-	return YES;
-}
-
-- (void)renderOnView:(UIView *)view {
-	if (!OSIPhone::get_singleton()) {
-		return;
-	}
-
-	OSIPhone::get_singleton()->iterate();
-}
+@property(nonatomic, readonly, strong) UIKitView *uikitView;
 
 @end
